--- examples/crewai/.env.example
@@ -1,3 +1,0 @@
-ARCADE_API_KEY=...
-ARCADE_USER_ID=...
-OPENAI_API_KEY=...
