--- examples/crewai/pyproject.toml
@@ -1,14 +1,0 @@
-[tool.poetry]
-name = "arcade_example_crewai"
-version = "0.1.0"
-description = "CrewAI examples with Arcade"
-authors = ["Arcade AI <dev@arcade-ai.com>"]
-
-[tool.poetry.dependencies]
-python = ">=3.10,<3.13"
-langchain_openai = "^0.2.14"
-crewai_arcade = {path = "../../contrib/crewai", develop = true}
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
