+++ examples/crewai/requirements.txt
@@ -1,0 +1,2 @@
+crewai-arcade>=0.1.0
+arcade-google>=0.2.0
